feedzirra working
added a rake method, model, controller and view so that rake fetch_feeds
will pass a url to the model to extract feeds and store in the db. Next
step integrate this into the main para rake file. Note Feedzirra needs
curl so see bookmark for fixing this on windows